refactor(svelte): move runtime/ and scene/ into feature directories
