+++ packages/svelte/src/lib/plot-area-brush.ts
@@ -1,4 +1,4 @@
-import { clamp, normalizedRect, type PanelBounds, type PlotRect } from "./plot-geometry.js";
+import { clamp, normalizedRect, type PanelBounds, type PlotRect } from "./scene/geometry.js";
 import { isBrushTooSmall } from "./plot-interval.js";
 
 /** Live brush corners; may be denormalized while dragging (x0>x1 or y0>y1). */
