+++ packages/svelte/src/lib/scene/Axis.svelte
