+++ packages/svelte/src/lib/scene/Batch.svelte
