+++ packages/svelte/src/lib/scene/Legend.svelte
