+++ packages/svelte/src/lib/scene/SceneView.svelte
