+++ packages/svelte/src/lib/scene/geometry.ts
