+++ packages/svelte/src/lib/scene/stratum-paint.ts
