+++ packages/svelte/tests/interval-state.test.ts
@@ -18,7 +18,7 @@ import type {
 } from "../src/lib/interaction.js";
 import { createPlotInteraction } from "../src/lib/interaction-controller.svelte.js";
 import { createIntervalState, type IntervalStateDeps } from "../src/lib/interval-state.svelte.js";
-import type { ContinuousZoomDomains } from "../src/lib/plot-geometry.js";
+import type { ContinuousZoomDomains } from "../src/lib/scene/geometry.js";
 import { buildIntervalSelection } from "../src/lib/plot-interval.js";
 import { withEffectRoot, withFlushedEffectRoot } from "./helpers/effect-root.svelte.js";
 import { modelFor } from "./helpers/model.js";
