+++ packages/svelte/tests/legend-focus-state.test.ts
@@ -12,7 +12,7 @@ import GGPlot from "../src/lib/GGPlot.svelte";
 import type { LegendFocusEvent, PlotInteractionScope } from "../src/lib/interaction.js";
 import { createPlotInteraction } from "../src/lib/interaction-controller.svelte.js";
 import { createLegendFocusState } from "../src/lib/legend-focus-state.svelte.js";
-import { createPlotRuntime } from "../src/lib/plot-runtime.svelte.js";
+import { createPlotRuntime } from "../src/lib/runtime/runtime.svelte.js";
 import { createSourceIdentityTracker } from "../src/lib/plot-semantic-keys.js";
 import {
   createSemanticKeyService,
